% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_env_code.R
\name{create_db_env_code}
\alias{create_db_env_code}
\title{Register db connection in db package}
\usage{
create_db_env_code(package_name, connection)
}
\arguments{
\item{package_name}{The name of the package being created.}

\item{connection}{The database connection to be stored and used throughout
the package.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generate a small R file to create an environment variable and
store the supplied connection when the db package is loaded.
}
