% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_execute_code.R
\name{create_execute_code}
\alias{create_execute_code}
\title{Generate code file for _execute functions}
\usage{
create_execute_code(package_name)
}
\arguments{
\item{package_name}{A string, the name of the package to create the file for.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generate the R file for a dbpkgr package's _query function. Which
gets the results of an SQL query against the connection.
}
