% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_schemas_code.R
\name{create_list_schemas_code}
\alias{create_list_schemas_code}
\title{Generate the code file for _list_schema functions}
\usage{
create_list_schemas_code(package_name)
}
\arguments{
\item{package_name}{}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generate the R file for a dbpkgr package's list_schemas function. Which
retrieves all the schema names available from the connection.
}
