% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_tables_code.R
\name{create_list_tables_code}
\alias{create_list_tables_code}
\title{Generate the code file for _list_table functions}
\usage{
create_list_tables_code(package_name)
}
\arguments{
\item{package_name}{A string, the name of the package to create the file for.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generate the R file for a dbpkgr package's list_tables function. Which
retrieves all the table names available from the connection.
}
