% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structure_code.R
\name{create_structure_code}
\alias{create_structure_code}
\title{Generate the code file for _structure functions}
\usage{
create_structure_code(package_name)
}
\arguments{
\item{package_name}{A string, the name of the package to create the file for.}
}
\value{
Returns \code{NULL} invisibly
}
\description{
Generate the R file for a dbpkgr package's structure function. Which
retrieves information_schema.tables.
}
