% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_get_db_structure.R
\name{dbp_db_structure}
\alias{dbp_db_structure}
\title{Get database structure}
\usage{
dbp_db_structure(connection)
}
\arguments{
\item{connection}{A database connection to get the structure for.}
}
\value{
A \verb{lazy tbl} with columns \code{catalog}, \code{table_schema}, \code{table_name}, \code{table_type}.
It will be a \code{tibble} if information_schema is not found.
}
\description{
Returns a table of information about the structure of the database.
Specfically, it attempts to read catalog, schema, table name and type from
\code{INFORMATION_SCHEMA.TABLES}. If \code{INFORMATION_SCHEMA} is not available, it
uses \code{\link[DBI:dbListTables]{DBI::dbListTables()}} and returns
\code{NA} for columns other than TABLE_NAME.
This function is useful for building queries for tables
in the format \verb{SELECT * FROM catalog.schema.table}.
}
\details{
The \verb{_structure()} functions generated for each package work the same,
without taking the connection argument. This function is exported for working
with connection information to enhance the package e.g. after calling
\code{\link[=dbp_init]{dbp_init()}} and before \code{\link[=dbp_load]{dbp_load()}}.
}
