% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\docType{data}
\name{dbp_env}
\alias{dbp_env}
\title{Environment to record active db packages}
\format{
An object of class \code{environment} of length 0.
}
\usage{
dbp_env
}
\description{
Environment to record active db packages
}
\keyword{datasets}
