% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_structure.R
\name{dbp_get_db_structure}
\alias{dbp_get_db_structure}
\title{Get database structure}
\usage{
dbp_get_db_structure(connection, schema = NULL)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{schema}{Optionally select results only for a specific schema.}
}
\value{
A data.frame with columns \code{TABLE_CATALOG}, \code{TABLE_SCHEMA}, \code{TABLE_TYPE}
and \code{TABLE_NAME}
}
\description{
Attempts to read catalog, schema, table type and table name from \code{INFORMATION_SCHEMA.TABLES}.
If \code{INFORMATION_SCHEMA} is not available, it uses \code{DBI::dbListTables} and returns
\code{NA} for catalog, schema and type. This function is useful for building queries for tables
in the format \verb{SELECT * FROM catalog.schema.table}.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

DBI::dbWriteTable(con, "mtcars", mtcars)

dbp_get_db_structure(con)
}
}
