% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\name{dbp_init}
\alias{dbp_init}
\title{Prepare package files based on a database connection}
\usage{
dbp_init(connection, package_name)
}
\arguments{
\item{connection}{A database connection object.}

\item{package_name}{A string to be used as the name of the temporary package and the prefix to its
functions. Must be a valid package name; lowercase no punctuation recommended.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Creates a folder under dbpkgr/, writing functions and roxygen docstrings.
Intended to be ready to be turned into a package with \code{\link[=dbp_load]{dbp_load()}}.

\code{\link[=dbp_package]{dbp_package()}} is the one-line way to get a package as the connection
interface up and running. However, as the automatically generated
documentation is quite limited, \code{\link[=dbp_init]{dbp_init()}} is provided to generate the
files without installing the package. Then you can make your own changes
to the package like adding more detailed roxygen documentation.
Once you're done,
you can pass the package name to \code{\link[=dbp_load]{dbp_load()}} to install and load
the package.
}
