% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_load.R
\name{dbp_load}
\alias{dbp_load}
\title{Create package based on a database connection}
\usage{
dbp_load(connection, package_name)
}
\arguments{
\item{connection}{A database connection object.}

\item{package_name}{A string to be used as the name of the temp package and the prefix to its
functions. Must be a valid package name; lowercase no punctuation recommended.}
}
\value{
Returns \code{NULL} invisbly.
}
\description{
Creates and loads a package with a number of functions
for interacting with a database connection.

This function will load the functions from R files created by \code{\link[=dbp_generate]{dbp_generate()}}.
If the files have not already been generated yet, this funciton will call
\code{\link[=dbp_generate]{dbp_generate()}} to create them first.
}
\section{{Generated functions}}{
 The following functions will be created for
the database:
\itemize{
\item{\code{_structure()}} - get a table of catalog, schema and table names.
\item{\code{_list_schemas()}} - get a character vector of schema names.
\item{\code{_list_tables()}} - get a character vector of table names.
\item{\code{_query()}} - get the results of SQL query on the connection.
\item{\code{_execute()}} - execute a SQL statement on the connection.
}

As well as functions for each of the tables in the database.
}

\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbWriteTable(con, "mtcars", mtcars)

dbp_load(con, "mydb")

# Get info about the database
mydb_list_tables()

# Each table is available as a lazy tbl
mydb_mtcars()

# And you can query with raw SQL
mydb_query("SELECT cyl, COUNT(cyl) AS count FROM mtcars GROUP BY cyl")
}
}
\seealso{
\code{\link[=dbp_generate]{dbp_generate()}} for generating R files.
}
