% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\name{dbp_package}
\alias{dbp_package}
\title{Create package based on a database connection}
\usage{
dbp_package(connection, package_name)
}
\arguments{
\item{connection}{A database connection object.}

\item{package_name}{A string to be used as the name of the temporary package and the prefix to its
functions. Must be a valid package name; lowercase no punctuation recommended.}
}
\value{
Returns \code{NULL} invisbly.
}
\description{
Generates a temporary package
}
