% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\name{dbp_package_path}
\alias{dbp_package_path}
\title{Get the path for a dbpkgr package}
\usage{
dbp_package_path(package_name)
}
\arguments{
\item{package_name}{The name of the package to locate}
}
\value{
A string path.
}
\description{
#' Get the path for a package already created via
\code{\link[=dbp_package]{dbp_package()}} or \code{\link[=dbp_generate]{dbp_generate()}}
}
