% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_set_connection().R
\name{dbp_set_connection}
\alias{dbp_set_connection}
\title{Load the connection to be used}
\usage{
dbp_set_connection(connection, package_name)
}
\arguments{
\item{connection}{A database connection object.}

\item{package_name}{A string to be used as the name of the temp package and
the prefix to its functions. Must be a valid package name; lowercase no
punctuation recommended.}
}
\value{
Returns \code{NULL} invisibly
}
\description{
To be included in the \code{on.load()} section of a package
where \code{\link[=dbp_generate]{dbp_generate()}} has been used to create R files for a connection.
}
