% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_unload.R
\name{dbp_unload}
\alias{dbp_unload}
\title{Uninstall and delete a dbpkgr db package}
\usage{
dbp_unload(package_name)
}
\arguments{
\item{package_name}{The name of the dbpkgr package to be deleted. This should
be a package created by \code{\link[=dbp_package]{dbp_package()}} or \code{\link[=dbp_generate]{dbp_generate()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Detaches the package from the search path and deletes the
package files. It is recommended to restart the R session afterwards.

You shouldn't need to call this function as it will be called automatically
on clean up if the package was created with argument \code{temp = TRUE}. It is
provided for more fine control of the cleanup, especially for when
\code{temp = FALSE}.
}
\seealso{
\code{\link[=dbp_package_path]{dbp_package_path()}} to locate the package files.
}
