% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_package.R
\name{remove_temp_package}
\alias{remove_temp_package}
\title{Uninstall and delete a temporary package}
\usage{
remove_temp_package(package_name)
}
\arguments{
\item{package_name}{The name of the package to be deleted.}

\item{path}{The path to the package files.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Uninstall and delete a temporary package
}
\examples{
\dontrun{
remove_temp_package(package_name, dbp_package_path(package_name))
}
}
