% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_R_file.R
\name{write_R_file}
\alias{write_R_file}
\title{Write R function to .R file}
\usage{
write_R_file(package_name, function_name, function_code)
}
\arguments{
\item{package_name}{A string, the name of the package to create the file for.}

\item{function_name}{A string, the name of the function for the R file.}

\item{function_code}{A character vector of text/code to be written to file.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Utility function to be called by all of the \verb{create_*_code} functions,
which will write their code to file with UTF-8 encoding.
}
\seealso{
\code{\link[=create_structure_code]{create_structure_code()}}, \code{\link[=create_list_tables_code]{create_list_tables_code()}},
\code{\link[=create_list_schemas_code]{create_list_schemas_code()}}
}
