% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_tables_code.R
\name{create_list_tables_code}
\alias{create_list_tables_code}
\title{Generate the code file for _structure functions}
\usage{
create_list_tables_code(package_name, h)
}
\arguments{
\item{package_name}{}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generate the R file for a dbpkgr package's structure function. Which
retrieves information_schema.tables.
}
