% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_get_db_structure.R
\name{dbp_get_db_structure}
\alias{dbp_get_db_structure}
\title{Get database structure}
\usage{
dbp_get_db_structure(connection)
}
\arguments{
\item{connection}{A database connection to get the structure for.}
}
\value{
A tibble with columns CATALOG, TABLE_SCHEMA, TABLE_NAME.
It will be a lazy table/remote connection if information_schema is found.
TABLE_TYPE.
}
\description{
Attempts to read catalog, schema, table name and type from \code{INFORMATION_SCHEMA.TABLES}.
If \code{INFORMATION_SCHEMA} is not available, it uses \code{\link[DBI:dbListTables]{DBI::dbListTables()}} and returns
\code{NA} for columns other than TABLE_NAME.
This function is useful for building queries for tables
in the format \verb{SELECT * FROM catalog.schema.table}.
}
\details{
The \verb{_structure()} functions generated for each package work the same,
without taking the connection argument. This function is exported for working
with connection information to enhance the package e.g. after calling
\code{\link[=dbp_init]{dbp_init()}} and before \code{\link[=dbp_load]{dbp_load()}}.
}
