% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\name{dbp_load}
\alias{dbp_load}
\title{Load temp package}
\usage{
dbp_load(package_name, temp = TRUE)
}
\arguments{
\item{package_name}{Name of a package already initialised with \code{\link[=dbp_init]{dbp_init()}}.
Must be a valid R package name.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Takes package files generated by \code{\link[=dbp_init]{dbp_init()}}, generates the
Rd files for them using \code{\link[roxygen2:roxygenize]{roxygen2::roxygenise()}}, installs and loads the
package, attaching it to the search path with the package_name provided.
}
