% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_package.R
\name{dbp_package}
\alias{dbp_package}
\title{Create package based on a database connection}
\usage{
dbp_package(connection, package_name, temp = TRUE)
}
\arguments{
\item{connection}{A database connection object.}

\item{package_name}{A string to be used as the name of the temp package and the prefix to its
functions. Must be a valid package name; lowercase no punctuation recommended.}

\item{temp}{Whether the the package should be detached and deleted at
the end of the R session. Set to \code{FALSE} to allow it to persist.}
}
\value{
Returns \code{NULL} invisbly.
}
\description{
Generates a temp package with a number of functions
for interacting with a database connection.
}
\examples{
\dontrun{
# demo db in memory
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbWriteTable(con, "mtcars", mtcars)

dbp_package(con, "mydb")

mydb_list_tables()
}
}
\seealso{
\code{\link[=dbp_init]{dbp_init()}} and \code{\link[=dbp_load]{dbp_load()}} which this function wraps.
}
